pullzone: get: rename CacheControlPublicMaxAgeOverride
The CacheControlPublicMaxAgeOverride field in the Get API and the
CacheControlBrowserMaxAgeOverride in the Update API refer to the same setting
but have different names.
This is confusing for the user.

Use the same names for the settings in both struct to prevent confusions.
CacheControlBrowserMaxAgeOverride is used as name because in the admin-panel the
setting is called "Browser Cache Expiration Time".