client: fix: data handed to newDeleteRequest() was missing from request
newDeleteRequest() had a function parameter called params that was ignored.
It was not possible to pass data that is send out as body of the http-request.

Change the newDeleteRequest() method to work like newPostRequest().
It accepts an interface{} that is, if not nil, marshaled to JSON and send out in
the body of the http-request.