client: fix: non-2xx response with empty body caused parsing error
When the client received an non 2xx response with an empty http-response body,
the returned HTTPError.Errors slice contained an
"could not parse body as APIError: unexpected end of JSON input" error.

This was misleading, the body was empty, there was nothing to parse.

This error was returned by json.Unmarshal because an empty buffer was passed to
it.
Check if the buffer is empty before calling json.Unmarshal() when processing
non-successful responses.