videolibrary/get: always pass includeAccessKey parameter
If the IncludeAccessKey parameter was false, it would not be passed as request
parameter to the get API endpoint. If the behavior on bunny.net would change
regarding if a missing includeAccessKey parameter means true or false, the
behavior of the bunny-go lib would change too.

By removing omitempty we guarantee consistent behavior, independent of a change
on bunny.net side regarding the default for a missing includeAccessKey
parameter.